# public/js/swagger.yaml
swagger: "2.0"
info:
  description: "Micro service for generating a docx file using a template and data"
  version: "1.0.0"
  title: "JBT Document Generator Service"
  termsOfService: "http://www.nukon.com.au/"
  contact:
    email: "info@nukon.com.au"
basePath: "/v1"
tags:
- name: "Document Generator"
  description: "Generate ms-docx files using templates and data"
schemes:
- "http"
- "https"
paths:
  /docx/generate:
    post:
      security:
      - basicAuth: []
      tags:
      - "Document Generator"
      summary: "Upload the document template and data"
      operationId: "postGenerate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Template and Data for producing a docx file"
        required: true
        schema:
          type: object
          properties:
            data:
              type: "object"
              description: "JSON object that will be used in the template"
              example:
                first_name: "Hipp "
                last_name: "Edgar"
                phone: "0652455478"
                description: "New Website"
                clients:
                - first_name: "John"
                  last_name: "Doe"
                  phone: "+44546546454"
                - first_name: "Jane"
                  last_name: "Doe"
                  phone: "+445476454"
            template:
              type: "string"
              description: "Base 64 encoded template docx file"
              example: "UEsDBBQABgAIAAAAIQAMTxYSrgEAAJYHAAATAAgCW0NvbnRlbnRfVHlwZXNdLnhtbCCiBAIooAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAC0lctu2zAQRfcF8g8Ct4FEJ4uiKCxnkccyMVAXXTPkyCIikQRn7MR/35FkC27hhM7DGwESNfceXhIz06uXtsnWENF6V4qLYiIycNob65al+L24y3+IDEk5oxrvoBQbQHE1O/s2XWwCYMbVDktRE4WfUqKuoVVY+ACOVyofW0X8GpcyKP2kliAvJ5PvUntH4CinTkPMpjdQqVVD2e0Lfx5IIjQosuvhx86rFCqExmpFTCrXzvznkm8dCq7s/8HaBjxnDCEPOnQrrxts6x44mmgNZHMV6V61jCGffTTSeL1qeQ/F2zIHOH1VWQ1jfacWoteAyJm3TTGutMq6Hf+rHEibBvDrKQbdI+3/WKpvqwo0H3Y6jxbzbtPFYLFXm3YDIg7pGJN/r2CeCh23ykmEZ3j8dTKKPfEkSOU9OU+nOPtROgkBzpyIYaecRKhBGYgXR9y7d16JQTjp34V1Ev9B+Ah/Rwv12MDXJ1Bx9+2lkxDEfR9k//z8SfQyb1lyh5xHH5DnSPzAtneDoqvOufUGiGRhHBWHWu3oyDPo0zlDN+UMmAPesp+qs78AAAD//wMAUEsDBBQABgAIAAAAIQAekRq38wAAAE4CAAALAAgCX3JlbHMvLnJlbHMgogQCKKAAAgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAjJLbSgNBDIbvBd9hyH032woi0tneSKF3IusDhJnsAXcOzKTavr2jILpQ217m9OfLT9abg5vUO6c8Bq9hWdWg2JtgR99reG23iwdQWchbmoJnDUfOsGlub9YvPJGUoTyMMaui4rOGQSQ+ImYzsKNchci+VLqQHEkJU4+RzBv1jKu6vsf0VwOamabaWQ1pZ+9AtcdYNl/WDl03Gn4KZu/Yy4kVyAdhb9kuYipsScZyjWop9SwabDDPJZ2RYqwKNuBpotX1RP9fi46FLAmhCYnP83x1nANaXg902aJ5x687HyFZLBZ9e/tDg7MvaD4BAAD//wMAUEsDBBQABgAIAAAAIQBsOHklWwEAAMIFAAAcAAgBd29yZC9fcmVscy9kb2N1bWVudC54bWwucmVscyCiBAEooAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAKyUTW/CMAyG75P2H6rc11C2sQ9RuGyTuG5MO4fUaau1SRV7H/z7BSogMAiXXCrZUf0+ee14PP1tm+QbLNZG5yxLBywBLU1R6zJn7/OXq3uWIAldiMZoyNkSkE0nlxfjV2gEuZ+wqjtMXBWNOauIukfOUVbQCkxNB9qdKGNbQS60Je+E/BQl8OFgMOLWr8EmezWTWZEzOyuc/nzZOeXztY1StYQnI79a0HREgitjCKyrKGwJlLM+zlIHyfhx/euY+ghEzlncEWwyIYS7mAgViMK3oI+DFgxP6Le1tAaNolSalvfur1y/228sR1o2gB81Vc9KgST/+odHIR+yExxHxuz8KPRQXiPWJCH5UUx50IV2s+hZscmEEG5jIqyG/4BhmwpBZG5HxHuS5FYF7NqwDvn6G5zJm5gMP7B4+/cyvWTIjYeYIMpomotF4xmyTW0g+N7mnfwBAAD//wMAUEsDBBQABgAIAAAAIQDEi4q8NwQAAIoRAAARAAAAd29yZC9kb2N1bWVudC54bWzsWF1v4jgUfV9p/wPyPrdJKKUUNYyGdqZbaVZCtDP7ODKJQ6I6tmUbGKbqf9977STA0GX52JfVzgshtu+5537YPnDz7lvJW3OmTSFFTKLzkLSYSGRaiGlMPj99POuRlrFUpJRLwWKyZIa8G/z6y82in8pkVjJhWwAhTH+hkpjk1qp+EJgkZyU152WRaGlkZs8TWQYyy4qEBQup06AdRqH7prRMmDHg75aKOTWkgiu30aRiAnxlUpfUmnOpp0FJ9fNMnQG6oraYFLywS8AOuzWMjMlMi35F6KwhhCZ9T6h61BZ6K4o3/HrLuyoDzmOgGQcOUpi8UKswjkWDEPOa0nxXEPOS1+sWKups+WtC3qcGd5ouoBQrwC24N5KReqOS+zxgfVdV/RExCncFU1UEIRoO+1DY9FkzKWkhGpjjUrOeXNgRp/T3vZYz1dBRxWloD+K5wcKNeQCzsOt23npo5iCAra37mFPFSKtM+g9TITWdcGC0iDot7EgygMNiItMlPlVr0YfDJh3HJAyj992Li0tSD92xjM64xZnL66jdvq1nRmtDCKLxww5ebgJ84KcbgaaT2QetAc8uFTAwinH+aKm2JHB2tSfE6w07F1HXkbMDTo39KmjJ9oL8INIK0BN5bf2LVLJCn8RlIwJIjUKOG2m/Dtvh8LpO7ngt7Zsz62l3+bMT7gKe8JEvwQSyu+QM0jqnPCZPWPh7XVTZgeV/wtQCso3OfE3ozEqfPJh+EOnWgvQbbeY/Sflcg4ed9w7G5WcsATVCVKyce1tN3ko+K/E2q+frAbdEyN+HcJ9VpIT84t8idIr9tBYchoIBT+EJGJ5q+7rd9QR/HL7aMeyRa0CrAavZBptZf4ILKAyrMZfuxH9WOU+qnDoeAOPTuspaMoTDEy50ZySx8K42eMNy2KOLvvkek477omgCu8RlJZFcguNVdTjL7LG2E2mtLI+11sU0P9I1Znk9/rxI2R8gEarSJj6Fh20GzKPyhto/OBXTOrhMn30cV63jpnHHVavRch8TID14+S3hBYgp83r4UXKIn5NOl32DeX3jDHKlcT35t+18hU35s51Xu/A/3M4bHQBb4h+v5kOa+JTb+mcPw0V96GXw/zySVQ4/eV9fgvpk3uhpPOe9YIDzDDcqKCJ8eIm064oZuQt+U+HuJ8Jwj0xAEuGvXidr4bwsGiUD6jUmX+/lkCbPXonUa0GvNiv9VehEoWGJHe1QIuO3mD6CEQqUSrsjpZxREBxjljEN/x2gGqwOca/nSUs7lvohrRRSJqXdz6DnA1HTx+8Ai5qv3e44wZLD98sefHfKVE3hmkfHUsF4xy9xXQuvvdBZeFWymkaFs5r1QcTkqu0We4rN63RmgTFqJecO1JIBb8YLKDRxw/CPCCo8mOGFYKPCJsDyoutmoTd8vl2b+J9DMFb/iTL4CwAA//8DAFBLAwQUAAYACAAAACEAS/7AzPsBAABMBgAAEgAAAHdvcmQvZm9vdG5vdGVzLnhtbMRUTW/bMAy9D9h/MHRPLLdBOhhxiiLBht6KdPsBqiwnQi1RkOR4+fej7MjJ5iDox2EXy5LIx/dIkYv736pO9sI6Cbog2ZSSRGgOpdTbgvz6+X3yjSTOM12yGrQoyEE4cr/8+mXR5hWA1+CFSxBDu7w1vCA7702epo7vhGJuqiS34KDyUw4qhaqSXKQt2DK9oRnt/owFLpzDgCum98yRI5wao4ERGmNVYBXzbgp2mypmXxszQXTDvHyRtfQHxKbzCAMFaazOj4QmA6HgkveEjkv0sCMVF+L2nmvgjRLadxFTK2rkANrtpDnJ+CgaStxFSvtrIvaqjnatyWajeIPkt9RgbVmLpTgBjuAuJKPsnVTd5yHU91TVfxEzek3MsSIBYuDwFgp/x4xMFJN6gPlYas6Tiy3xmff9w0JjBjpGfg7tUb8OWKEz38GMzrvOO5fm3gUwat3nHTOCJIrnj1sNlr3UyKjNZkl4kWR5Ni2SNvcHg7dOGGaZB0vwSJYFmWSdocEtjqNyUxBK7+a3Dw+rYNEdrUXFmtqPb57C0eqO3qJxiGaebFicYRxfM7qzygtsbBxvbV7LkK+b2bDZNIEuazyQdLlIB/ceI/Lsr/AsGHTfOAEv6uOgvdRNNxGeI0bUSnuWUddY0OZ/SL1I+ZpszETMgVv+AQAA//8DAFBLAwQUAAYACAAAACEADA8HxPsBAABGBgAAEQAAAHdvcmQvZW5kbm90ZXMueG1sxFXbbtswDH0fsH8w9J5YboN0MOIURYIWfSvS7QNUWY6FWqIgyfHy96N8SzcHQS8Pe7EtiTw8hxTp1e1vVUUHYZ0EnZFkTkkkNIdc6n1Gfv28n/0gkfNM56wCLTJyFI7crr9/WzWp0LkGL1yEENqljeEZKb03aRw7XgrF3FxJbsFB4eccVAxFIbmIG7B5fEUT2n4ZC1w4h/E2TB+YIz2cmqKBERpjFWAV824Odh8rZl9rM0N0w7x8kZX0R8SmywEGMlJbnfaEZiOh4JJ2hPrX4GEnKs7E7Ty3wGsltG8jxlZUyAG0K6U5yfgsGkosB0qHSyIOqhrsGpMsJvFGye+pwdayBktxApzAnUlG3jmpqstDqO+pqv8iJvSSmL4iAWLk8B4Kf8ccmCgm9QjzudS8TS52xFfu94OF2ox0jPwa2qN+HbFCY36AGV22nfdWmvsQwKR1n0tmBIkUTx/3Gix7qZBRkyyicCPJ+jQsoib1R4OHThhmmQdLcEvmGZklrZ3BJQ6jfJcRSm+W13d3m2DRbm1FwerKT0+ewtbmhl6jcQhmnmx4OcM4XmZ0Z4UX2Nc43Jq0kiFdV4txsasDW1Z7IPF6FY/uHcbAszvCvWDQPvv5d04dB+2lrttx8DwgDEppx3FQNZWz+x9Cz1K+IBrTMPwA1n8AAAD//wMAUEsDBBQABgAIAAAAIQB7HN0roA8AAMUjAAAQAAAAd29yZC9oZWFkZXIxLnhtbOxaaW+r2Jb93lL/hyhP6i9+CfOUqtwnJuMBbGOMMW61SsxgM5kZP/V/7wNOcpN7U69vDWqppUKyDWdYZ5+9z7QW/vkfXRLfNV5RRln6fI88wvd3XupkbpQGz/f6bvpA39+VlZW6Vpyl3vN975X3//jy7//2c/sUusUdqJ2WT23uPN+HVZU/QVDphF5ilY9J5BRZmfnVo5MlUOb7keNBbVa4EAoj8HiXF5njlSVoirfSxirvX+CS79Gy3EtBW35WJFZVPmZFACVWca7zB4CeW1VkR3FU9QAbJl9hsuf7ukifXgx6eDNoqPJ0M+jl57VG8V0vPmn3VlPInDrx0mpsESq8GNiQpWUY5V+78XvRQBfDV5Oaf9WJJolfy7U5gn/X3luXfyQGQmG1IBRfAb+D+8QZ7q1SEt/8MMT3a1S/RUTgf9WZl4gMEG82/IgJH9t8tSSxovQN5ve55r1zwWT4I+NbKrI6fzMnj/4Y2jw9v2ENc/I3WAaT48x737XyNwF8N3W10Mq9+7vEeZoHaVZYdgwsahH8bhiR91/AOpHftU9gfXG3z/cwjJEYRYA15SVJ8HyrjqshhyIxluVfczZDEs8QMIzcQDbFiKVVfeyB2o0VP9/PPMv1invoy88QaOZWYixW3O7TbFNkmX/Lf0nLI6f68nPzVA52V33u3UXu8/0vHQyuXyoURu/vnAyMwTK6go6gCAnDfx+/7++ypzIHpo5lwJoTPt8nf49vmcWtYHz76TxgM2iiKrKzd3fKorQczAblo+pmb/M0ANwFheVGYBUZjcnOz/fV0IqTpannVINxz/cFuBt7+M7kN/vf216iMAFW71utv33o0EvreVZGwyr1ZNllFteV9xNYQ4MofYg9v3p6QOFHLK9e06osf3ogHwmQ0kZuFT4hDPnIDI+hFwVh9UThj9TweH2IUtfrnlACIUkUw8ifmqh8WY+fxtsYNFRmD21h5Q+jKU/lpbaKl9QB/CH3Cgf44Qn+WtKNhn3H8W7WMYNlryBvWYORn1UpRgs/r2NnVQU2gFu1V5c8hFkRXbO0suJ33gHtfVLg4bbaN95T5XU3q95KgZ20ipxfBXnN/gRidPK3fhh989bcLVqjG24x+ArzktU8DCY9ALeB/jwB7wzDKfA716ossJ2LpcBxKssF7JwF11zl0eAaaVAAHtYiB753Q7oBrxo73cbg4+9ncXs8rFw5DSt7oW1VnRIUjpv+x9869icGPjmtPzcWOkmpbBCT5hLWYD6bSdx51y3lo9QROylsTsQalWBta6yyK0TRac20TneeKtIqIC+1epHJLTsCwiVhklFBNBUTC8ii5YR5u1YWy1k+i/OCZJa+Ozv1nZYUkeab14mupcbGpNjaX80td0kq4kLpGDYXoEwYAdfxrsiK8IDGUzGKE2eNbt1VdeKJTpgyOV/vepPjCDsibRUv+2yBBYyhwj3YPeRrQCxd/hAv5VPjGDSDjYCavr4c2vRKCEWQzqrFrluXqi8u5ztKSARN562upbaXlKZo6lgmW4Xx0H2y8bgVXl/MZlcZ1KmJTgGzgvoRcFEePEe+wJeJIZAJDdwvQFCrsN8Hi8UhbQYhQ7AW6hAsGXzYA7aI7ZnSyOkiHgHt2fp8VqaX9soqYT0RGnOz9Q+S1XbRaRdNTrDU5LNSMKNEsQzY7FG/QBV8wylpYes2pDJ4DPlQqETIZXuzcD9Hy3VF6ovAnRwjIfB9mfbO03iq2c0ezjJctPnwKnW4z0hb10aJmVKa4SXmW4mIC/nC2NHx3Bnr6GKOgChqRrvCQs+mUiMxUqkCBkP2JsHa1lzBHUIHNk6wrgqycuIElVMO3cFTldrC6W4l7rETmTG+p16LDIlGwLXBIvg0mai0hVj0hGtqyIVcxUYXJlFHGzuUoHOV4s5F0iGvwHKnaflf8/MIOFfZfOf2uNCCibJxBODn9eDs42xxlVG9t+FtY/YEbht9qcNIgwJfNypk+h0HHyat6RxNyzEv9cF0uBEQU32Vspf2vEw6RUz1Bb9HmbqSe+Ns2bIbboJI13ZUv9nLYJhPMydnJ0uZrVKHWWqXXTxf7NzK31EVVN7G4RLO+ePUbIW9hTJ8IDQuSfuF7h14pSBFUSbVhEdXrFJc9sxpiedRyVHWPjKDo2KkWwRv+4UHWbiaBuh8tBBM9m3sGb2ksDKBz2RxIzSYp50T/Oq7R3kmI/DhyGtk7ob2delviKvldoa9SnIkLWKfRlibCntc0TtLHgHTy6w2pO2ZYva+Lh47oZXI9XoV8HVAdsYhCaRrQmyZcpY2yXFt8XiwxeIopTfIaTLxhb3Y7636shOVTDmNgJNMEubnpbZUnEw76mcp7jJlI29No3UwWppnNAEHWMlCJ0XlKYcprxWBX2GlMVTeIPr0up5HrjNf8IZKjYA+u+QqN5iogWmt0YnmX4K5elAdxBfpWd0LzqY7Ip5BG5WlOxutmjbGpW+UwtisHerQz2e9ap6z3vDCDhoBt4k3CztutyI7QjinU3nVFoWvLC2xluKY2u+P8E4+dWCSdGqhXg15pWCU3kYTl1ymXH+E8/WKs1Wm1AlxBFzQeYfNcTNZ7szgFER+C+0WtSyrljKDiFpjba/NvdkhTeNF5BOZ7qyLibHcIqejP2329lG80rnhUqQhGSPgHNnK897Yedvmejkzjo0IxS7Rj4dTRjBmYceuTjSmBje5j802h/5wcjfDwIcgWmCnnMy2qs6yJgumz23FZnmRpdxCX1fD0gWWJBVMrBbcSdve6Zmzjbm1k+zRcf8Q+01rULzC9hJFbFAhB57I1uK0EoyVftiNFuYxjJ0MZI6JixOaFdLlmLM15flrZ7aKTxADNQS9sXMjBSP0pDkrmjjzwX5Loggf455P+X42n5+3AlJo3gioK4rPxcGBwTyit1qfZJtNggQL4SKpkoogpymcHLWwcHlb8ee+WOGbJjUTmskNuNmecUzWe93I4tRIbjNlw2hJyzhWP9kkhLCj/bBDUfxAFivKKcOGvvAasSBTpEW3tKeaW8fsDhJPkQWLT9cTjImMVi1MO1cv8G09JAzJc6JzAieGFVTBNbkE4nExk9GWmm4GX24giG05VWtFXlbZtwAMvq8yBT83OCgkhIPvxy6LQ6VPrl/d2IfIqvOK/XA4uO03I+C3m84HbJ4GmxHngyiXTk8P0S4d5FPAcU0dAb9dWD8C4mB4cWcnVZrjbtFsJeY0Dp9PrRSqERCtVBIbu10OW+M4aD+ASmfxDVRiMBtsl240rOAd8Ksu3cqCsxErXEfAIcEdBju4xoT7Oz+KYw+QBX+QRIaj/Xg/HMaHU5eddQNDqTq748FBEpxmvyM/YJgQuPhKcbbvyA/N4RgClIsbLfqW/Iycpvryz4HmAFTwfSNDA7kRiwLUup36y9yLY62yipExvJSqvsRWWf2SWon3Q/XF1B34Bqj9StLe8baxh1/+++73muJHxZ9ty4dODTzwO7f/Ouf8c9wOWFzy3uu/6rc8BBraB3uBud8GcUD7kRh8wPmN/f7okffDbYz8DzL2/zcow7R5Py2hdzMW3I/s+6YhjOLA6/z6TVLCSMWPIL53rnc3rAfeHZASBsngnb4AODrO/AGOjmHUIzmQ7hfiPpJ05HEgui8kHSceSfD0wtFx/JEGT+8pOsGgJP7HKfoLAf2Lt3/G2z93zl9k/i8y/6a8/EXm/w/JPCeH9aLdqwOZn4Nj4K+T+b0S14DMq83WzzYdTyPUYo9S+rSHZ5eVaNzIvCvOdof0cqr3IRB+iKWSW04xmahlqKfcYZ55kYRfZ+XV7qxaLgrrPHDnTKlIIC3l560lnGX7oMcQVIST8WhZkoLIoeZUQblYUyENxhuERAYpMT4LoW7Otwgn1rUMnvOpvtgdF4DwWf1xBosneYdfHEGXvZlG7YTTC81Lq6pMWzpNnGq9EnF8taAaGbvMJmeS1kqg90aXeOFxyWIR2bxUyRfMriW7mGoVejoZe6J0WlRgRH23uh2vYS9mXUfn55C7huD93ppQbNBmJyHVJS051+f2qCvdGdKxlV3qZiIfAwdru5xuJps4hY4+thIjQl4vZakcu9yFMtOe59u5nqiHI3s+9fFFQiYBYaBCdZlSQoca6zJFrirWrmG6tPk8cSFZWxtB0auatJ6nc4q9WjQU3lhZmcBQAHM+JRDsvL7oF3QzUXhFRO2pHpIiLbF+CHv7qC7bg92n4UQAwgthX/fyRJlqLoxjV3y5FBNnv9JGC8PJfkNJHYTDayUEkT1PTgoqh6QyzR0L3WjzneVhXDhjgs4+Fk2x3/XXKMggYyFZiytXTU5ov6XN89rK6BGQDODZHKsajDzw++Vaa0TI0xEoouqdbMiaIs73W8HiTl0+4XdCxJ0Kc7uINyeBFrqewSVyom5tLmIr2w5GwENgFVhDnuvKsayOJO2jTy5L/WBTZiDEIclBsIpfQJTbwmF7vRISbgF5DEVEakzkmykTx+u8UvDSO7cjYGpjQNkT1gfKTmW663M+z7PlzrfPS4u+nBZtZKCJu8zZDj+tHUsJCN3CoWDRKWdm03tFRzIXSHVGwjoC3lirIk6BHsCywcClQlYI2Krd+MMDrw6zcVAOXujeR2amnOhqDXRMdeu7tyhDcwXR1p3FwXNPOMa6rhrz3daDwaiUFlky7+gpeoYELOH0coutYIpKY7cSBaA0IY0miTWmi7vp9HQU6dvAbpum84664CCkOp2DQPt8NZP6giQrbbqSFWZpJQHKHHl+euDnu/WEmgWYYLhwuayuFgi3QCqIs5zB1NEZu2ycpJkjYO2kD5f11mRC87g+ScXh4tlXZyXO3fwCkReg2RjaBt2T/qFeMNAxJ7pVbecHXCSqQwShAo6o0GUE1FkjLalEdaWO1WTlLEQnfhngO7bgFnC7MyVtunFqbt/7dHOZjUrAq4DMsRHMvvf9LSiyJawX8mQIAD4EgFOG2+8voMsAMRc+GgR8QLaEI+2vB+2VMt+kiBvgiyA06BHrDRxCjAnAeH3QI/iRk3/Abob097K1PVu9aRvfAWq4HTlTfdBr+4HZg9Hz7VUPI+t7aXYg9iI/An6rmqxZHmOZUbFavClWH2D5ZmjsBrptXJTojwez9iTktnwBdX4KXqMMF3CgytJDN+US+Hq4np/HVm+KgZPFGfhjQRuCd393/4mBfz/816t+8G0Wgr5lteNLNvCyMa/Gd4k/qDB8pGR/vsLwLdW96QGuVzpFlA/vFj+w0v+V3b7ICR8q3ajs76Jst6qg1U8Y+GvSBsQChqc8RWH4q97y3k8fPfieGo+6wo+9hv7UEJ6Csa8vuN+3+ZLzmxsAzYA/4Xz5HwAAAP//AwBQSwMEFAAGAAgAAAAhAFNbFMB3AgAAkggAABAAAAB3b3JkL2Zvb3RlcjEueG1stFZbb9owFH6ftP8Q+R3iQMW6iFCtUKa+VaN7nkziJFZ9k23I0LT/vuOEQIEOhbLxAAh8vsv5fA6M734KHqypsUzJBEV9jAIqU5UxWSTo+/O8d4sC64jMCFeSJmhDLbqbfPwwruLcmQCqpY0rnSaodE7HYWjTkgpi+4KlRlmVu36qRKjynKU0rJTJwgGOcP1OG5VSa4FqSuSaWLSFE6doSlMJXLkygjjbV6YIBTEvK90DdE0cWzLO3Aaw8aiFUQlaGRlvBfV2gnxJ3AjavrQV5sTFG7xN5UylK0GlqxlDQzloUNKWTO9tvBcNLJatpPU5E2vB23OVjm5O+HaWu2QwM6SCKPaAJ3BvNCNrigRv+uDz3ad6jBjhc2a2iXiInYYuEg45WyWCMLmDeV9rXjcXhuGa+/3VqJXeydHsOrRH+bLD8jN5gTI8qifvtTV7EcDJ6C5KoikKRBo/FlIZsuSgqIpuAn8j0QT2hA6qGPZL9i1BGE8/4eGXKWo/mtGcrLjz38wH0cNg1lQ8mbpw4TacwtE14QmaK+WoQeFkHAJmc8K0OE8wuhgP7oef57iGMM2BVHHlDzUIcwyPLcL2gJv88niuRm1YjVL5g/FVbqPBjNWU84UjxvlS8NKSesrb+5thNLqQkhPrfkgiaCfqB5kdEV/h9vcRZWPZkaU3VW+xNX2GzQnbtfATVMWEs0L6VZegFJ4hAviQU5JR6LmEH4W2oQe9+ZvEkhUlAJYOQOpQNtBcVbUYTWr/PpSOtDkz/yeZjvxXpWN8W8+G4xPq0NvCENHhvruJLiH+owsF6/9wfDza/go3Eg59+nmu5w/+UUz+AAAA//8DAFBLAwQUAAYACAAAACEAlrWt4pYGAABQGwAAFQAAAHdvcmQvdGhlbWUvdGhlbWUxLnhtbOxZT2/bNhS/D9h3IHRvYyd2Ggd1itixmy1NG8Ruhx5piZbYUKJA0kl9G9rjgAHDumGHFdhth2FbgRbYpfs02TpsHdCvsEdSksVYXpI22IqtPiQS+eP7/x4fqavX7scMHRIhKU/aXv1yzUMk8XlAk7Dt3R72L615SCqcBJjxhLS9KZHetY3337uK11VEYoJgfSLXcduLlErXl5akD8NYXuYpSWBuzEWMFbyKcCkQ+AjoxmxpuVZbXYoxTTyU4BjI3hqPqU/QUJP0NnLiPQaviZJ6wGdioEkTZ4XBBgd1jZBT2WUCHWLW9oBPwI+G5L7yEMNSwUTbq5mft7RxdQmvZ4uYWrC2tK5vftm6bEFwsGx4inBUMK33G60rWwV9A2BqHtfr9bq9ekHPALDvg6ZWljLNRn+t3slplkD2cZ52t9asNVx8if7KnMytTqfTbGWyWKIGZB8bc/i12mpjc9nBG5DFN+fwjc5mt7vq4A3I4lfn8P0rrdWGizegiNHkYA6tHdrvZ9QLyJiz7Ur4GsDXahl8hoJoKKJLsxjzRC2KtRjf46IPAA1kWNEEqWlKxtiHKO7ieCQo1gzwOsGlGTvky7khzQtJX9BUtb0PUwwZMaP36vn3r54/RccPnh0/+On44cPjBz9aQs6qbZyE5VUvv/3sz8cfoz+efvPy0RfVeFnG//rDJ7/8/Hk1ENJnJs6LL5/89uzJi68+/f27RxXwTYFHZfiQxkSim+QI7fMYFDNWcSUnI3G+FcMI0/KKzSSUOMGaSwX9nooc9M0pZpl3HDk6xLXgHQHlowp4fXLPEXgQiYmiFZx3otgB7nLOOlxUWmFH8yqZeThJwmrmYlLG7WN8WMW7ixPHv71JCnUzD0tH8W5EHDH3GE4UDklCFNJz/ICQCu3uUurYdZf6gks+VuguRR1MK00ypCMnmmaLtmkMfplW6Qz+dmyzewd1OKvSeoscukjICswqhB8S5pjxOp4oHFeRHOKYlQ1+A6uoSsjBVPhlXE8q8HRIGEe9gEhZteaWAH1LTt/BULEq3b7LprGLFIoeVNG8gTkvI7f4QTfCcVqFHdAkKmM/kAcQohjtcVUF3+Vuhuh38ANOFrr7DiWOu0+vBrdp6Ig0CxA9MxHal1CqnQoc0+TvyjGjUI9tDFxcOYYC+OLrxxWR9bYW4k3Yk6oyYftE+V2EO1l0u1wE9O2vuVt4kuwRCPP5jeddyX1Xcr3/fMldlM9nLbSz2gplV/cNtik2LXK8sEMeU8YGasrIDWmaZAn7RNCHQb3OnA5JcWJKI3jM6rqDCwU2a5Dg6iOqokGEU2iw654mEsqMdChRyiUc7MxwJW2NhyZd2WNhUx8YbD2QWO3ywA6v6OH8XFCQMbtNaA6fOaMVTeCszFauZERB7ddhVtdCnZlb3YhmSp3DrVAZfDivGgwW1oQGBEHbAlZehfO5Zg0HE8xIoO1u997cLcYLF+kiGeGAZD7Ses/7qG6clMeKuQmA2KnwkT7knWK1EreWJvsG3M7ipDK7xgJ2uffexEt5BM+8pPP2RDqypJycLEFHba/VXG56yMdp2xvDmRYe4xS8LnXPh1kIF0O+EjbsT01mk+Uzb7ZyxdwkqMM1hbX7nMJOHUiFVFtYRjY0zFQWAizRnKz8y00w60UpYCP9NaRYWYNg+NekADu6riXjMfFV2dmlEW07+5qVUj5RRAyi4AiN2ETsY3C/DlXQJ6ASriZMRdAvcI+mrW2m3OKcJV359srg7DhmaYSzcqtTNM9kCzd5XMhg3krigW6Vshvlzq+KSfkLUqUcxv8zVfR+AjcFK4H2gA/XuAIjna9tjwsVcahCaUT9voDGwdQOiBa4i4VpCCq4TDb/BTnU/23OWRomreHAp/ZpiASF/UhFgpA9KEsm+k4hVs/2LkuSZYRMRJXElakVe0QOCRvqGriq93YPRRDqpppkZcDgTsaf+55l0CjUTU4535waUuy9Ngf+6c7HJjMo5dZh09Dk9i9ErNhV7XqzPN97y4roiVmb1cizApiVtoJWlvavKcI5t1pbseY0Xm7mwoEX5zWGwaIhSuG+B+k/sP9R4TP7ZUJvqEO+D7UVwYcGTQzCBqL6km08kC6QdnAEjZMdtMGkSVnTZq2Ttlq+WV9wp1vwPWFsLdlZ/H1OYxfNmcvOycWLNHZmYcfWdmyhqcGzJ1MUhsb5QcY4xnzSKn914qN74OgtuN+fMCVNMME3JYGh9RyYPIDktxzN0o2/AAAA//8DAFBLAwQUAAYACAAAACEAiITafFkEAABCCwAAEQAAAHdvcmQvc2V0dGluZ3MueG1svFZLb+M2EL4X6H8wdK5jvSzZQpyFX2lTJG1QZy+9URJtEyuSAklZcX99h6QYxbGyWPTQk8j5Zj4OZ4Yzuv3ySqvRCQtJOFt4wY3vjTAreEnYYeF9fbkfz7yRVIiVqOIML7wzlt6Xu59/um0ziZUCNTkCCiYzWiy8o1J1NpnI4ogpkje8xgzAPRcUKdiKw4Qi8a2pxwWnNVIkJxVR50no+4nX0fCF1wiWdRRjSgrBJd8rbZLx/Z4UuPs4C/Ej51rLDS8aipkyJ04ErsAHzuSR1NKx0f/KBlc8OpLT9y5xopXTawP/e5rddVsuyjeLH3FPG9SCF1hKSBCt7HUpIuyNJoiviN5CfQOhntizJ5oKzAPfrHrPZXVlP5Btm8VHkgskbJqhALQXtMgeDowLlFdQVG0Qe3dQUf9wTkdtVmNRQJKgHH3fm2hAYMpP+BnKlDNUPTBbUZC6d+gGKbxk5Quh2EghAHy/UyAFSlnjqjI1XVQYgQNtdhCIQjUuPCsxNiXeo6ZSLyjfKV6D0gnBPdOwc+N4ro+YmZr5G16Dw+Nwat0sjkigQmGxq1EBp605U4JXTq/kf3C1hsoXkBhrcSzF7ohqvLEHy7tbnkkt6DyRo1OGXyEUuCQKXmJNSopeF17oTwPN0GlX6MwbdaGrMa1cX0hHJVIIzLXt5MIY9m127U6b7TlXjCv8LHQm3A7uRMqFNzZeXIlNvIDPaVtbzMqeqNt84LmUOpoLQ9s6tC92tbNtCIgYolBMVtq1lideYp3sRpCrev203rWBSXxg4+ROMpUoSQlZgpKExV8QGqfr+9tlMtusbV412iPBLJ5uu0h9QJZJFHXVc4mEYTyNt0Ns4SxJks0gsorm9+7JXHgQJVE6nQ3ZxJt0vR1km86DMBy8TxIE89kwEkdAN3ROkqZBMBiDZB2vgkHf0iRaLgfPSdNwvkqHzpmt4ihIBpF734+XQ8jcD/3VfAhZp3407ME6jTdhPGizSX1oaKahXOZ0E0WzYPA+n9fOdp34q2iI7T4MtqHJHLw0fRDUJc30GNLPza7uoQONqK3mNaK5IGj0pAcVeEezXHxbEebwHENbxe+RXZM7cDy2gKSoqu6hyznAlBvNSiJraGKGtnpC4tDzBnHcKYnPAOh2v78x6t6Pxa+CN7U9sxWofmAliN2hPSVh6pFQJ5dNvnNWDEbOO6hh5Z8noQknfZDaTMGfCtZRekTs4F4sZuOvO60KL78SO/03g59QXUNLB5X8ECy8ihyOKtCdQsGuhL8as8kPYYeFBoOdxswGFfpmoN0ttIJdgla36GWRk0W9DGa21Yt72dTJpr0scTL4q4KmfoaBCsPvG3RAt9TyPa8q3uLyNydceFciGwQzuB5YUTUlhpooeSEfmB6t0sTIwMtGcTfbnkmhGhhyBjXj7H+eb6bPDMy3D760WYkLAhW9O9O8n9M39tYVkWqHaxjpiguIlxkJv5gK6n9+7/4FAAD//wMAUEsDBBQABgAIAAAAIQC7L7iidQEAAD8DAAAUAAAAd29yZC93ZWJTZXR0aW5ncy54bWyUUsFuwjAMvU/aP1S5Q9qpIFTRIiGENGmnjX1AmqZttCSOkkAHXz/TwsbGDnCKY/u9+PllvvjUKtoJ5yWYnCTjmETCcKikaXLyvlmPZiTygZmKKTAiJ3vhyaJ4fJh3WSfKNxECdvoIWYzPNM9JG4LNKPW8FZr5MVhhsFiD0yzg1TVUM/extSMO2rIgS6lk2NOnOJ6SE427hQXqWnKxAr7VwoQeT51QyAjGt9L6M1t3C1sHrrIOuPAe9Wg18GkmzTdNkl4RackdeKjDGMXQYSJ6pEJ4EveRViTSPHtuDDhWKtxgl6SkwPVVcudPZ9RlssrJbDpJZ2k6mfb1Eqr9Su6wtmMKrSH02I3LexF1OGfj7+yrbNp/0huw171LCAH0nzzOs6zc8Y3wgzFoOsFGf8gJfg0MLOMooo85KECv2DbAMIa6mOw+ZPlrovuw7lL5PVDam9CLHsJiPpy9L2CD1PIg1uCWDjovHBqA9YtvX3wBAAD//wMAUEsDBBQABgAIAAAAIQC81LsrwQcAAGE9AAAaAAAAd29yZC9zdHlsZXNXaXRoRWZmZWN0cy54bWy0m99T2zgQx99v5v4Hj98hP6DkyjTtUCgtM22PEph7VmyFaLAtn+0QuL/+VpKtGDu2d2P3qcSx9rOrXX1XUOnDp5cwcJ55kgoZzd3J8dh1eORJX0SPc/fh/vroL9dJMxb5LJARn7uvPHU/ffzzjw/b8zR7DXjqgIEoPd/G3txdZ1l8Phql3pqHLD0OhZfIVK6yY0+GI7laCY+PtjLxR9PxZKx/ihPp8TQF2iWLnlnq5ubCujUZ8whYK5mELEuPZfI4ClnytImPwHrMMrEUgchewfb4rDAj5+4mic5zh46sQ2rIuXEo/6cYkdSi2MM1I6+ktwl5lGniKOEB+CCjdC3iXRiHWoMQ14VLz21BPIdB8d42npzWeDZkTA6uEraFVOwM1sztmQzfDAoDMw8qv7usVi1Oxm3B5BlRJqwPGBfeMgtPQiYia+awqSlPLqyHPvX9NZGb2LoTi37WbqIna0stS4Jn4zO98sqhpSQDtaW7WLOYu07ond88RjJhywA82k5OHVWR7keQCl96V3zFNkGWqo/JbZJ/zD/pf65llKXO9pylnhD3ICFgJRRg8NtFlAoXvuEszS5Swcpffsmfqe/X6sXyl3akl2Ylg5+FL9yRgqb/wbBnFszd6bR4cqmcePMsYNFj8WyVHF3flZ2Zu/bREuzOXZYcLS6UsZGOtPi3FHH8Jn74pF2JmQeLDzhslXHQIRAyxQmESvB0BqJmPtxt1PyyTSZziDYAsLJZ+FiZdJAnEKuFEW34lq++S++J+4sMvpi7mgUPH25uEyETUNK5+/69YsLDBQ/FN+H7XPWI/NlDtBY+/2fNo4eU+7vnv661QucWPbmJMnD/bKYLIUj9Ly8ej5VSgumIqST/VANAxiAdJY52aCN23pgHFap++G+BnJgc7qWsOVNdzdH+t4J01JveoKmKqByAtkvy9aS/idP+Jt71N6GLt99czPp7AXuZvhkxtVGqSnxSM+mZ4ivPw8n7lpJVI2pV1DmiVjSdI2o10jmiVhKdI2oV0DmilvDOEbX8do6opbN1hMe0cFWr6ETPBmph34ssgFbZoXSTnlKXtxrnliXsMWHx2lG9tep2m1guNssM56qW08PFcpElUu04O2YEurNaugdr8pcwXrNUwMa8C9Rz6u/V7sf5mgjYwXag3pniq8WkNyZ7W9htwDy+loHPE+eev5iMEsb/lM7C7DI6neuZ1u/icZ05sDFULbcTdtYw6c0zYex/F6meg9ZuftYQSpdxVA7PGuqy2fgP7otNWEwNYjdyZvSckOYKQrvYPkWnKkX11dUZhUoAJgTTLughaPsI/01zodtXOcb4b1rRgfYR/pvGdaB9XR/t+SUrzRX8ZcVBLa8Zee1eykAmq01QrIFOeZiRV7BF4EIgL2JrHyUSM/IKfiOfzoXnwW9umDol52KnowQKOR2GohcbPhZyUiqyNyFERE5QhTUlsPppLQFEFt07/izU34GpzUCrtN1rdi7nk4YZgBaE2kP/2sisew89bdA8LOUmgj+XpNzB0U4aVh6WlteT6XeEHPdrfARQvw5IAPVrhQRQQ30073lsT8RD+jdHAossy7aL6bJDK/OMrMwWRGsBA/VNxP6rYfU210K9byIo5ATV+yaCQs5OpZfZvolgDdY3EayGrtGco7KmUoIi980yyO4EEBENI94I0DDijQANI94IUH/x7oYMJ94IFlkbrKaWxRsB0q9QftW3oLJ4I0BkbTBql//NqOh72kr7L7cDiDeCQk5QXbwRFHJ2msQbwdKvUCqhwrJSh2ANI94I0DDijQANI94I0DDijQANI94IUH/x7oYMJ94IFlkbrKaWxRsBIsuDBZXFGwHSr1C0Ya9461X/28UbQSEnqC7eCAo5OxVBtZtUBIucoArLijeCpV+hFEPO0sVNCWoY8UZENIx4I0DDiDcCNIx4I0D9xbsbMpx4I1hkbbCaWhZvBIgsDxZUFm8EiKwNe8VbL8bfLt4ICjlBdfFGUMjZqQiq1TkEi5ygCsuKN4Kl66W3eCNA+pVDQZSIhhFvRETDiDcCNIx4I0D9xbsbMpx4I1hkbbCaWhZvBIgsDxZUFm8EiKwNe8Vbr5HfLt4ICjlBdfFGUMjZqQiqFW8Ei5ygCstKHYI1jHgjQLowe4s3AqRfOQCkVxElTcOINyKiYcQbAeov3t2Q4cQbwSJrg9XUsngjQGR5sKCyeCNAZG1Q52zhvCj6eOqkoQiw5wyKUw1o4LQhSVhgHuAdX/EELhby7tMhPYFFhARiQ3lgQ/ws5ZODO9h90lAgaJRYBkLqI92v+pRO6SLCyazlJsH935fON3MBpjZOl9Tbkzdwe6h8XUhfT1IXh8DP7DWGKztxcbJcWYMLQupqV34FSF8LvYELQfm1HjVY3fOBF/Wlqvyx/n/bnAo/A1EPrKO8NbA8uBHVgsoPvNszSPq4exXccCpeO7K7klG4mZ+O3+2hzHtvzmi2+p2pk+AtPuuT4q1z5OhXTFbrDsLlLO1Sl4eQsmVgrpjBDzeRDxFu89tZJpn+CzOm4PtLHgQ/mL6Qlsm4+dWArzLz7WSsO2DF1FJmmQybxyf6gLj2ZJ8BKIeyM+ajCqK5TqJNuORJfty8sSRV59A30d6WpDnr2lAK2Jne+Vb8lH78HwAA//8DAFBLAwQUAAYACAAAACEAD+3a5fMBAAC7BQAAEgAAAHdvcmQvZm9udFRhYmxlLnhtbJyTTW7bMBCF9wV6B4H7WJSspIkROWjdGOimi8I9AE1TFlH+CBzaqm/fISkrBQwjVihAkGbIh5mPb55f/mqVHYUDaU1NihklmTDc7qTZ1+T3Zn33SDLwzOyYskbU5CSAvCw/f3ruF401HjI8b2CheU1a77tFngNvhWYws50wmGys08zjr9vnmrk/h+6OW90xL7dSSX/KS0ofyCDjblGxTSO5+G75QQvj4/ncCYWK1kArOzir9beo9dbtOme5AMCetUp6mkkzyhTVhZCW3FmwjZ9hM3mqKA9SeLyg8Usrkmm++LE31rGtQnZ9UZHlAC7rF4ZpDK6YklsnY6JjxoIoMHdkqia0pGt6j+/wVHQe3iQPCrxlDoQfN9IUbpiW6nSOQi8BUqKTnrfn+JE5GQpKKZB7TBxgS2vySnGV6zVJkaImFQa+rsZIiUWlVQx75mMEnYOFRZ24pXiKOhhBneFUrDNP1rkgsZFaQPZT9Nkvq5m5QqSkD0jiHnkEMvNJRFzUjQQnEClfx/6xkxW28uWxmg/9vxF5ep9I0plAhLVY8RUQ3xBEMEVAUX3cGsb6jTuIzakTU8AMF3rG8P8VJ/O8gYnGQINdtwql0U63g1kxjTNzjUywRuISrDJtaD5mkcuhodVomikk3h+aYXpg+Q8AAP//AwBQSwMEFAAGAAgAAAAhABnyjblMAQAAbQIAABEACAFkb2NQcm9wcy9jb3JlLnhtbCCiBAEooAABAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAHySUUvDMBSF3wX/Q8l7m6TDWUPbgcqeHAhWFN9CcrcFm7Qkcd3+vWm71Q2Gj/feky/n3CRf7HUd7cA61ZgC0YSgCIxopDKbAr1XyzhDkfPcSF43Bgp0AIcW5e1NLlomGguvtmnBegUuCiTjmGgLtPW+ZRg7sQXNXRIUJgzXjdXch9JucMvFN98ATgmZYw2eS+457oFxOxHRESnFhGx/bD0ApMBQgwbjHaYJxX9aD1a7qweGyZlSK39oQ6aj3XO2FONwUu+dmoRd1yXdbLAR/FP8uXp5G6LGyvS7EoDKXAomLHDf2DLH50VYXM2dX4UdrxXIx0OYX+mFloWd6t+lpINiKgN7iDJeADIK5tgY5TT5mD09V0tUBn/3MclieleROUszRshXb+fifG92bOijqX+JNIsJjWlakQc2Sy+JJ8CY6fKDlL8AAAD//wMAUEsDBBQABgAIAAAAIQDJCKHYgQgAAN1DAAAPAAAAd29yZC9zdHlsZXMueG1s3Fzbcts4DH3fmf0Hjd5TXzdpMnU7idNsMtOmbezsPssSbXMqiV5RbpJ+/YKgJOtGi5JD7yUvtikKBwCBQ8oG8u7Dc+BbP0jEKQsn9uBN37ZI6DKPhquJ/Ti/OXlrWzx2Qs/xWUgm9gvh9of3v/7y7umCxy8+4RYICPlF4E7sdRxvLno97q5J4PA3bENCuLhkUeDE8DFa9QIn+r7dnLgs2DgxXVCfxi+9Yb9/aidiIh0pbLmkLrlm7jYgYYz39yLig0QW8jXd8FTak460JxZ5m4i5hHMwOvClvMChYSZmMK4ICqgbMc6W8Rswpic16glRcPugj+8C37YC9+JuFbLIWfjgvKfB2H4PnvOYe02WztaPufgYfY2Sj8knfLlhYcytpwuHu5RO7Knj00VEbRghDo8vOXUKg+vLkBenuXxiz2kAS3RPnqwHFjih3ROifSdcgZgfjj+xSXjyOCsKzYYW1AOJTnQyuxQ39lDT9DWn8SbTX84qmQcLA8s0k9ECxpPlJ+Z+J94shgsTGyIOBx/vvkaURRARE/v8PBmckYDeUs8jIjjTieGaeuTPNQkfOfF2499uMNISiS7bhvHEHp6eoct97n18dslGxAjghU4A0PfiBlgliOUcDiq0pTtt5EAJFQf/SiEHiWfrUNbEEelkof57gXDG9mCgYaNFrwQ0OhbQWADllwQNaOX93w4XASx1qBZnh4sAQj5UC8yubu6c09gnQoO9gSzXq3Mgz7aLWAtHLmpnnCsfiMiak+e40SIM9e6pOYsjFq4aUTBzu6N8DDZrh1PY/xrW50CcudjOrN8j6jUiKZhG7kN1bPmJrtaxNVsjZzaKH3aU/4ny5jUfKtzUpLyWY4aKiFIL/0w8ug1S12hsJ0NFGmpDNG8kQ0UGNkKIBdAxQXJuJcP15GvoLwm5vXyxxjr6S7buKF9Df0nlHeU3798jxdlH7f9rOOJbWuk1ap27U+azaLn10xxopIdR6wzOIPRMaJ3EmXwtkhi1zuACfVqXrgtHb404HbXOYwmEaayP0jqbJQommz5K15xONh1r0MJvXfM7xRq2wOqa65Jr9YHGrZP+gfyg4hsEkZDV44uaLFCzr07krCJns25OZ0U6AKrWs8C3LYubD69DhZ91Ue5CeN7lxNJDUzGsLlpyKkBPtgimcWvyTYDkxtcimA4COppFI5MWZXuiSZBsZ0mpRZsw28dChtWO/w8AarcFtAeq3Tebz1/j1geM6r5pDiXvNBMo5SeSNLyPgJURg0Gs4r5pGugYFhXJ26BFCHQ0izLyNmFRhbxNgGSEWiJvo1h58jYKZJqHasm7+eF23PrprUre5lDyTjOBoiLvI2BlxGAQq0jepoGOYVGRvA1aVCRv00AZeZsAqpC3CRAVeRvFypO3USDTPFRL3vigv/8nJsV3Afu+bBC/LOT9Zg4l7zQTKCryPgJWRnUGsYrkbRroGBYVydugRUXyNg2UkbcJoAp5mwBRkbdRLNMklBllmodqyRt/S9hP3q/wc4M5lLzTTKCoyPsIWBnVGcQqkrdpoGNYVCRvgxYVyds0UEbeJoAq5G0CJOO50tcmRrHy5G0UyDQP1ZI3/kS7n7w7/oqb95s5lLzTTKCoyPsIWBnVGcQqkrdpoGNYVCRvgxYVyds0UEbeJoAq5G0CREXeRrFMk1BmlGkeEjWyUISpXfI5eKWqBm1AVWmmbmFDYuADWZIIWlJIc3WIotJAFzCt22iBiL+aVAtedBGvGIPyY61K5xF+x3cAFF34lGGZzQtW6eT6G0Zne7oX5l+m1q3sYKjchyFVrLyBJpF8v4doN8E+IXBJ/LKBnotNvtoHekFEF0zSw4ET76CjI+nLEDeLRg24F3tVkmH8hi9BxffQvOSlc/r9weXpCMrZ8Ap0pqAGG8cVHRjQTLOMCfQYQcuRMManoqVpeIaMIz48bEWbjrONWdLwkgiADh2U8zOFgdJaBOA/p6JHB9WTY7kuGbSn6gF3DS5wQQ+hgsIDSXdNVholmoGwjyXvj2SSlc2ycBoqtqvuT9VLWlR2Rzs5r1A6CkOgv0LvWFR979EZq8L3Lp2FU6TnqgpC04/0adaLU68hRNLClwsCb+5CsfbQ6YUrKmPMe3akKLg+Jb7/2cHli9lGPdUny1heHfSxqK8kasHimAXq+yOsW0dN6gSAW/PKyI/CCLW/w22wIBHE7R6f3zOxoVUi457JElzpwCzVQXuMVF1Pq3UrZHGWt4InIKzLeSsaoGAYlVk40LP1RbRgVVIaEvB7Oi4lTSFTmqKlSImIUeSD6Vl/dDmVYpJ0hkjGxjt4TQFFnazMyA2Dvrnx6dt+EkK7ObjEYjFwyvnoFKeIpUR58Ablq71WyHx3yyGgZqItrsx/OevLvpSXrJ1nSg6tJQ70StW9Ta5V+/FV+bA+lm4Y1GpWY2kph9vEkpS085iaef6PsZSzvhxL8tKhsSSl/EtiqZBhO16Ck0ck4r9CTbfZlbqIUifTATE0Oh2d/fZW+ivJI1f0NaRM1Ie/mxtBMzH0URN8oID246Ki23S26FmGXRm0B9pBcWr6qU+0K8f3GQuxC64cIck12SJX56H8kSxPMDmhOonXdCgB0wpnvJvh4OPwWvpQdcbLn/BkRTso2HTCi0pd13NnDX3TYjGS/urdALZXy8voGZ4dDgenUrP84VCONa9RIYD3bBFlB+9buUMzPIclDVNHf2kh1auWRP4/7m/YupN+VrERi4ePuRgRT+8VssArsu2xLhfwcj4hqm5KzuBFt5z3h/2r5OybP9lewX8TgH/MIFZWnlzxxJQkPEQk/wnnFBGaHB5v4AyBAY9UsnuEgWccea7tdG925u10d3oi7nQzhf9m4JHbEs1pWy1v/6Pb7ZClsA559/+XHiNS/ufv/wYAAP//AwBQSwMEFAAGAAgAAAAhAICAOJjOAQAAwQMAABAACAFkb2NQcm9wcy9hcHAueG1sIKIEASigAAEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAnFPBbtswDL0P2D8YvjdKgqLoAkZF0WLoYVsDxG3PqkwnwmRJkFij2dePshvH2XaaT++RNPX0SMHNe2uLDmMy3q3LxWxeFui0r43brcun6uvFdVkkUq5W1jtclwdM5Y38/Ak20QeMZDAV3MKldbknCishkt5jq9KM044zjY+tIqZxJ3zTGI33Xr+16Egs5/Mrge+Ersb6IowNy6HjqqP/bVp7nfWl5+oQWLCECttgFaH8keVYEGMAKk/KVqZFOefwSGCjdpjkAsQA4MXHmvkSxIDgbq+i0sTeyasvICYUbkOwRitiU+V3o6NPvqHisb9+kX8HMS0BtmSL+i0aOmQVUwrfjBt0DIB1RbWLKuw/xI0MtlpZvON7y0bZhCBOAXhAlWe6UYblQkerDjX5WCTzi6e6LItXlTC7tS47FY1yxK7lsoH02IZEUVaGLPfm3MB7OC2bYnOZHeRaBueFOTho4MS5uv6E9NjwTekfYhdTsb2GQepEzgSOZ/zRld38mZ5C5e/zWnwYdh6cTPjF0H4blOZJXPOATrOeJGDLC4E1D+/Y7hSAB7Y22nwm/+t2WB9r/k7k7XkeXqRcXM7m/PXrcozxRo5PRf4GAAD//wMAUEsBAi0AFAAGAAgAAAAhAAxPFhKuAQAAlgcAABMAAAAAAAAAAAAAAAAAAAAAAFtDb250ZW50X1R5cGVzXS54bWxQSwECLQAUAAYACAAAACEAHpEat/MAAABOAgAACwAAAAAAAAAAAAAAAADnAwAAX3JlbHMvLnJlbHNQSwECLQAUAAYACAAAACEAbDh5JVsBAADCBQAAHAAAAAAAAAAAAAAAAAALBwAAd29yZC9fcmVscy9kb2N1bWVudC54bWwucmVsc1BLAQItABQABgAIAAAAIQDEi4q8NwQAAIoRAAARAAAAAAAAAAAAAAAAAKgJAAB3b3JkL2RvY3VtZW50LnhtbFBLAQItABQABgAIAAAAIQBL/sDM+wEAAEwGAAASAAAAAAAAAAAAAAAAAA4OAAB3b3JkL2Zvb3Rub3Rlcy54bWxQSwECLQAUAAYACAAAACEADA8HxPsBAABGBgAAEQAAAAAAAAAAAAAAAAA5EAAAd29yZC9lbmRub3Rlcy54bWxQSwECLQAUAAYACAAAACEAexzdK6APAADFIwAAEAAAAAAAAAAAAAAAAABjEgAAd29yZC9oZWFkZXIxLnhtbFBLAQItABQABgAIAAAAIQBTWxTAdwIAAJIIAAAQAAAAAAAAAAAAAAAAADEiAAB3b3JkL2Zvb3RlcjEueG1sUEsBAi0AFAAGAAgAAAAhAJa1reKWBgAAUBsAABUAAAAAAAAAAAAAAAAA1iQAAHdvcmQvdGhlbWUvdGhlbWUxLnhtbFBLAQItABQABgAIAAAAIQCIhNp8WQQAAEILAAARAAAAAAAAAAAAAAAAAJ8rAAB3b3JkL3NldHRpbmdzLnhtbFBLAQItABQABgAIAAAAIQC7L7iidQEAAD8DAAAUAAAAAAAAAAAAAAAAACcwAAB3b3JkL3dlYlNldHRpbmdzLnhtbFBLAQItABQABgAIAAAAIQC81LsrwQcAAGE9AAAaAAAAAAAAAAAAAAAAAM4xAAB3b3JkL3N0eWxlc1dpdGhFZmZlY3RzLnhtbFBLAQItABQABgAIAAAAIQAP7drl8wEAALsFAAASAAAAAAAAAAAAAAAAAMc5AAB3b3JkL2ZvbnRUYWJsZS54bWxQSwECLQAUAAYACAAAACEAGfKNuUwBAABtAgAAEQAAAAAAAAAAAAAAAADqOwAAZG9jUHJvcHMvY29yZS54bWxQSwECLQAUAAYACAAAACEAyQih2IEIAADdQwAADwAAAAAAAAAAAAAAAABtPgAAd29yZC9zdHlsZXMueG1sUEsBAi0AFAAGAAgAAAAhAICAOJjOAQAAwQMAABAAAAAAAAAAAAAAAAAAG0cAAGRvY1Byb3BzL2FwcC54bWxQSwUGAAAAABAAEAAEBAAAH0oAAAAA"
      responses:
        201:
          description: "unique ID for the generated request"
        401:
          description: "unauthorized"
  /docx/generate/{id}:
    get:
      security:
      - basicAuth: []
      tags:
      - "Document Generator"
      summary: "Download the document"
      operationId: "getGenerate"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of the generated document"
        required: true
      responses:
        200:
          description: "successful operation"
        401:
          description: "unauthorized"
securityDefinitions:
  basicAuth:
    type: basic